fix uninstall dependencies power shell script to work with log file

+++ utils/scripts/windows/uninstall-web-pro-dependencies.ps1
@@ -7,15 +7,15 @@ $version = ""
 $infos = @()
 $no_choco_installed=$False # assume chocolatey is installed
 
-$CHOCO_UNINSTALLED_TEST = 1
-
 Write-Host ""
 Write-Host "$PSCommandPath" -ForegroundColor Blue
 Write-Host "Preparing for uninstallation of $install_name and related packages ..." -ForegroundColor White
 
 Init-Log "$PSScriptRoot\..\..\..\logs\uninstall-web-pro-dependencies.log" | Out-Null
 
-$TEST_FORCE_INSTALL = 0
+$CHOCO_UNINSTALLED_TEST = 0
+$TEST_FAKE_UNINSTALL = 0
+
 
 # Set Execution Policy
 Write-Host "Setting execution policy: Set-ExecutionPolicy Bypass -Scope Process -Force"
@@ -34,8 +34,9 @@ try {
     exit $ERROR_SETTING_EXECUTION_POLICY
 }
 
+# Set security protocol
+Write-Host "Setting security protocol -bor 3072"
 try {
-    Write-Host "Setting security protocol -bor 3072"
     [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072
 } catch {
     $err = "Error setting security protocol!"
@@ -66,7 +67,9 @@ if (-not (Get-Command choco -ErrorAction SilentlyContinue) -or $CHOCO_UNINSTALLE
 Write-Host "Uninstalling $install_name packages ..."
 try {
     $version = choco --version
-    $output = & { choco uninstall all -y }
+    if (!$TEST_FAKE_UNINSTALL) {
+        $output = & { choco uninstall all -y }
+    }
 } catch {
     $err = "Error uninstalling Chocolatey packages"
     Write-Error $_.ToString()
@@ -80,9 +83,12 @@ try {
     exit $ERROR_UNINSTALLING_CHOCOLATEY_PACKAGES
 }
 
+Write-Host "Uninstalling $install_name $version ..."
 $choco_path = "C:\ProgramData\chocolatey"
 try {
-     $output = & { Remove-Item -Recurse -Force $choco_path }
+    if (!$TEST_FAKE_UNINSTALL) {
+        $output = & { Remove-Item -Recurse -Force $choco_path }
+    }
 } catch {
     Write-Error $_.ToString()
     $err = "Error removing Chocolatey $version"
@@ -90,24 +96,30 @@ try {
         error = $err
         exception = $_.Exception.Message
         exception_full = $_.ToString()
-        $errors = @("${err}: ${choco_path}")
+        errors = @("${err}: ${choco_path}")
     }
     Log-Progress -message ($result | ConvertTo-Json)
     exit $ERROR_UNINSTALLING_CHOCOLATEY_REMOVE_DIR
 }
 
 # Try removing any other folders related to Chocolatey but just output warnings when failed
+Write-Host "Removing directories related to $install_name $version ..."
 try {
-    $output = & { Remove-Item -Recurse -Force C:\ProgramData\ChocolateyHttpCache }
+    if (!$TEST_FAKE_UNINSTALL) {
+        $output = & { Remove-Item -Recurse -Force C:\ProgramData\ChocolateyHttpCache }
+    }
     $infos += "Removed directories related to chocolatey $version."
 } catch {
     # Do nothing
 }
 
 # Clean up environment variables
+Write-Host "Cleaning up environment variables related to $install_name $version ..."
 try {
-    [System.Environment]::SetEnvironmentVariable("ChocolateyInstall", $null, [System.EnvironmentVariableTarget]::Machine)
-    [System.Environment]::SetEnvironmentVariable("ChocolateyHome", $null, [System.EnvironmentVariableTarget]::Machine)
+    if (!$TEST_FAKE_UNINSTALL) {
+        [System.Environment]::SetEnvironmentVariable("ChocolateyInstall", $null, [System.EnvironmentVariableTarget]::Machine)
+        [System.Environment]::SetEnvironmentVariable("ChocolateyHome", $null, [System.EnvironmentVariableTarget]::Machine)
+    }
     $infos += "Removed environment variables related to Chocolatey $version"
 } catch {
     Write-Error $_.ToString()
@@ -116,7 +128,7 @@ try {
         error = $err
         exception = $_.Exception.Message
         exception_full = $_.ToString()
-        $errors = @($err)
+        errors = @($err)
     }
     Log-Progress -message ($result | ConvertTo-Json)
     exit $ERROR_UNINSTALLING_CHOCOLATEY
